zlepseni prace s api, uprava struktuy
