+++ frontend/craco.config.js
@@ -1,0 +1,24 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const path = require('path');
+
+module.exports = {
+  webpack: {
+    alias: {
+      '@': path.resolve(__dirname, '../'),
+      '@frontend': path.resolve(__dirname, 'src'),
+      '@components': path.resolve(__dirname, 'src/components'),
+      '@utils': path.resolve(__dirname, 'src/utils'),
+      '@hooks': path.resolve(__dirname, 'src/hooks'),
+      '@types': path.resolve(__dirname, 'src/types'),
+      '@assets': path.resolve(__dirname, 'src/assets'),
+      '@styles': path.resolve(__dirname, 'src/styles'),
+      '@services': path.resolve(__dirname, 'src/services'),
+      '@core': path.resolve(__dirname, '../core'),
+      '@shared': path.resolve(__dirname, '../shared'),
+    },
+    configure: (webpackConfig) => {
+      webpackConfig.resolve.modules = [path.resolve(__dirname, 'src'), 'node_modules'];
+      return webpackConfig;
+    },
+  },
+};
