+++ frontend/src/children-dashboard/pages/Dashboard.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Box, Card, CardHeader, CardBody, Heading } from '@chakra-ui/react';
 import ChildrenTable from '../components/ChildrenTable';
-import api from '../../services/api';
+import api from '@frontend/services/apiConfig';
 import { texts } from '../../texts';
 import { useLanguage } from '../../shared/contexts/LanguageContext';
