+++ frontend/src/types/child.ts
@@ -4,10 +4,25 @@ export interface Child {
   surname: string;
   date_of_birth: string;
   notes: string;
-  parent_firstname?: string;
-  parent_surname?: string;
-  parent_email?: string;
+  age: number;
+  parent_id: number;
+  parent_firstname: string;
+  parent_surname: string;
+  parent_email: string;
   parent_contact?: string;
+  parent?: string; // computed field
   confirmed?: boolean;
-  status?: 'accepted' | 'denied' | 'pending'; // Add this property
+  status?: 'accepted' | 'denied' | 'pending';
+}
+
+export interface CreateChildData {
+  firstname: string;
+  surname: string;
+  date_of_birth: string;
+  parent_id: number;
+  notes?: string;
+}
+
+export interface UpdateChildData extends CreateChildData {
+  id: number;
 }
